server/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
threadpoolctl==3.4.0
tqdm==4.66.2
tzdata==2024.1
Werkzeug==3.0.1
